// ts/react/common/types/bedev2Types.ts
export enum TTreatmentType {
	Carousel = "Carousel",
	AvatarCarousel = "AvatarCarousel",
	SortlessGrid = "SortlessGrid",
	FriendCarousel = "FriendCarousel",
	InterestGrid = "InterestGrid",
	Pills = "Pills",
	Sdui = "sdui",
	SongCarousel = "SongCarousel",
}
// ...
export type TTopicLayoutData = {
	componentType?: TComponentType;
	playerCountStyle?: TPlayerCountStyle;
	playButtonStyle?: TPlayButtonStyle;
	hoverStyle?: THoverStyle;
	infoText?: string;
	hideSeeAll?: "true" | "false";
	navigationRootPlaceId?: string;
	isSponsoredFooterAllowed?: "true" | "false";
	linkPath?: string;
	subtitleLinkPath?: string;
	endTimestamp?: string;
	countdownString?: string;
	backgroundImageAssetId?: string;
};
// ...
export type TSongData = {
	assetId: number;
	title: string;
	artist: string;
	albumArtAssetId: number;
	isPrivate: boolean;
	duration: number;
};
// ...
export type TExploreApiSongsSort = {
	topic: string;
	topicId: number;
	songs: TSongData[];
	treatmentType: TTreatmentType.SongCarousel;
	nextPageToken: string;
	sortId: string;
	contentType: string;
	subtitle?: string;
	topicLayoutData?: TTopicLayoutData;
};

export type TSongSort = TExploreApiSongsSort;